Split hierarchy.rs into sibling.rs and descendants.rs
